+++ lab_10_02_02/inc/process.h
@@ -9,7 +9,7 @@
 #include "consts.h"
 #include "input_output.h"
 
-double cnt_val(node_t *head, double val, int cnt_coefficients);
+int cnt_val(node_t *head, double val, int cnt_coefficients);
 void cnt_derivative(node_t *head, int cnt_coefficients);
 void cnt_sum_of_polynomials(node_t *head, node_t *add_head, node_t **result, int cnt_coefficients, int cnt_coefficients_for_mode_sum);
 void separator_coefficients(node_t *head, node_t **head_ch, node_t **head_nch, int diff);
